fix: move button component path
